# shared/types/index.ts
// Generated types will be added by the contract generator (UoW-02 onwards).
// For now we hand-define the only contract type that exists: HealthResponse.
export interface HealthResponse {
  status: 'ok';
  ts: string;
}